Intial Commit
This is my git repository recreation of Eli Criffield's Sipie.
The project appears to be a little slow and in a non working state.
My goal is to bring this project up to date.

DISCLAIMER: I've never worked on a public project before, I can't
garuntee I will stick with it as Im a real flakey guy.