It works. Added more dirty regex hacks.
Planning on cleaning up the regex later.
Maybe I'll play with this a bit in the future.

If anyone runs into some HTML Parser Error please enable the
print data in the 2nd last line fo Sipie/Factory run it like
sipie.py &> sipie.log and email me sipie.log, or feel free to
add a regex and send the line to me, and I'll add it.