Rewrite of all Copyright Headers
I only put my name into the copyright headers, that doesn't mean I am
claiming all copyrights. It is just not my place to add others.

If you have worked on any of this and wish for your name to be in the
copyright headers feel free to do so.